feat:修改初始sql文件、修改相关中间表实体其它实体文件

+++ server/src/modules/system/role/entities/role-dept.entity.ts
@@ -7,9 +7,9 @@ import { Entity, PrimaryColumn } from 'typeorm';
   comment: '角色和部门关联表',
 })
 export class SysRoleWithDeptEntity {
-  @PrimaryColumn({ type: 'int', name: 'role_id', default: 0, comment: '角色ID' })
+  @PrimaryColumn({ type: 'bigint', name: 'role_id', default: 0, comment: '角色ID' })
   roleId: number;
 
-  @PrimaryColumn({ type: 'int', name: 'dept_id', default: 0, comment: '部门ID' })
+  @PrimaryColumn({ type: 'bigint', name: 'dept_id', default: 0, comment: '部门ID' })
   deptId: number;
 }
