+++ server/src/modules/system/role/entities/role-menu.entity.ts
@@ -7,9 +7,9 @@ import { Entity, PrimaryColumn } from 'typeorm';
   comment: '角色和菜单关联表',
 })
 export class SysRoleWithMenuEntity {
-  @PrimaryColumn({ type: 'int', name: 'role_id', default: 0, comment: '角色ID' })
+  @PrimaryColumn({ type: 'bigint', name: 'role_id', default: 0, comment: '角色ID' })
   roleId: number;
 
-  @PrimaryColumn({ type: 'int', name: 'menu_id', default: 0, comment: '菜单ID' })
+  @PrimaryColumn({ type: 'bigint', name: 'menu_id', default: 0, comment: '菜单ID' })
   menuId: number;
 }
